feat(notification): Added Notification Mailbox Model and migration scripts
